feat: Enhance Python server with TLS, Poetry, and Refactoring
This commit introduces a comprehensive set of improvements to the Python server, enhancing its security, maintainability, and functionality.

Key changes include:

1.  **TLS Encryption:** The WebSocket and TCP servers now support TLS (`wss://` and secure TCP) to encrypt all network traffic. This is configurable and can be enabled for secure deployments. Self-signed certificates can be generated for local development.

2.  **Poetry Dependency Management:** Migrated the project from `requirements.txt` to Poetry. This provides deterministic builds via a `poetry.lock` file and simplifies the development setup. The README has been updated with the new installation instructions.

3.  **Major Code Refactoring:** The main `gesture_server.py` has been refactored for improved modularity. Core components like the `SystemController`, `GestureExecutor`, and data models have been moved into a new `Server/Python/core` directory, making the codebase cleaner and easier to maintain.

4.  **New 'Volume Control' Gesture:** Added a new gesture to control the system's master volume. This demonstrates the extensibility of the refactored architecture.
