+++ Server/Python/pyproject.toml
@@ -1,0 +1,30 @@
+[tool.poetry]
+name = "gesture-control-pro-server"
+version = "1.0.0"
+description = "Python server for GestureControlPro"
+authors = ["Jules"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.11"
+uvloop = "^0.21.0"
+websockets = "^15.0.1"
+pyautogui = "^0.9.54"
+pyperclip = "^1.9.0"
+aiohttp = "^3.12.15"
+pyyaml = "^6.0.2"
+numpy = "^2.3.2"
+scikit-learn = "^1.7.1"
+translators = "^6.0.1"
+websocket-client = "^1.8.0"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.4.1"
+pytest-asyncio = "^1.1.0"
+black = "^25.1.0"
+flake8 = "^7.3.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
