--- Server/Python/requirements.txt
@@ -1,31 +1,0 @@
-# Serveur GestureControl Pro - Dépendances Python
-# Version optimisée pour performance maximale
-
-# Core dependencies
-uvloop
-websockets
-pyautogui
-pyperclip
-aiohttp
-pyyaml
-numpy
-scikit-learn
-translators
-
-# Network & Communication
-websocket-client
-
-# Optional: Computer Vision (pour extensions futures)
-opencv-python
-mediapipe
-
-# Optional: Machine Learning (pour prédiction avancée)
-# tensorflow-lite is not available on PyPI for Python 3.12 on this platform yet.
-# It is also not used in the current version of the server.
-# tensorflow-lite
-
-# Development & Testing
-pytest
-pytest-asyncio
-black
-flake8
